+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## next [????.??.??]
-* Depend on `text-short-0.1` or later.
-* Depend on `base-4.7` or later.
-* Add `TextShow` instance for `ShortText` in `TextShow.Data.ShortText'
+* Add a `TextShow` instance for `ShortText` (from the `text-short` package) in
+  `TextShow.Data.ShortText`.
+* Drop support for `base-4.5` and `base-4.6` (GHC 7.4 and 7.6, respectively).
 
 ## 3.7 [2018.10.07]
 * Remove `hoopl` dependency. Instances are no longer provided for data types